+++ core-service-lib-integration/src/test/java/org/x2b/studi/core/TestIntegrationService.java
@@ -15,7 +15,6 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.test.web.servlet.setup.DefaultMockMvcBuilder;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
-import org.x2b.study.core.IntegrationTestService;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @WebAppConfiguration
