+++ core-service-lib/src/main/java/org/x2b/studi/core/security/User.java
@@ -1,0 +1,7 @@
+package org.x2b.studi.core.security;
+
+import java.util.UUID;
+
+public interface User {
+    UUID getUUID();
+}
