--- core-service-lib/src/main/java/org/x2b/study/core/security/User.java
@@ -1,7 +1,0 @@
-package org.x2b.study.core.security;
-
-import java.util.UUID;
-
-public interface User {
-    UUID getUUID();
-}
